Added gitter and build badges in README
It appears not everyone can easily locate the link to the Gitter chat in the
README, hence I thought about adding a gitter badge at the top to improve
discoverability. As I was doing this, I've also added a travis badge.

Furthermore, I've added a link to the contributors gitter channel.