Avoid redefining sbt update task for dev runtime configs. Fix #74
As explained in the related ticket, adding the `Defaults.baseClasspaths`
settings to the `DevRuntime` ivy configuration, was preventing the update cache
to properly work, and effectively cause full dependency resolution to be
triggered every time the `runAll` task was evaluated.

The fix implemented in this commit is to simply scope dependencies needed only
during development to the `DevRuntime` configuration, instead of adding tasks to
the configuration.