Reworked URL parameter handling (#99)
Instead of representing parameters extracted from the URL as a single
object that gets passed to a ServiceCall, and having to define
IdSerializers to handle that, we now handle the parameters as individual
method call parameters that get passed directly to the service interface
method. This greatly simplifies working with URL parameters, both
conceptually, and also from an implementation perspective, since
IdSerializers could be very difficult to implement.