Renamed all "with" methods to something better
It was really inconsistent under what circumstances the API used "with",
and when it used "withFoo" - it usually had to do with overloading, and
this would limit us with API migration going forward.  So I've made it
consistent, everything is named more descriptively, I think this also
makes the code easier to read and discover.