Support for facilitating creating a read-side Cassandra table
As highlighted in lagom/activator-lagom-java-chirper#24, making sure that a
read-side table is (eventually) created can be challenging. In particular, users
need to ensure that the create table statement is retried if Cassandra isn't
available.

This commit addresses the problem by encapsulating the retry logic in a new API
class: `CassandraReadSideTableCreator`. Both the number of attempts and the
interval between each attempt is configurable.