Really fix Kafka spec
The old spec was not just checking that Lagom reresumed consuming
messages when an error occured, it was checking that Kafka redelivered
them. The problem is, it seems sometimes Kafka can take a long time to
redeliver them, which meant things were failing in CI.

This new test simply tests that when the stream fails, Lagom will reresume
consuming it.