Avoid stream materialization on requests with no body (#274)
This change yields a 4x performance improvement for a service call that
does nothing but returns a String (hello world). The problem is that
Akka streams materialization is very slow. When there is no request
body (eg, for GET requests), there is no need to materialize a stream to
handle it, and so this change avoids materialization in that case.