Fix off-by-one error in the JDBC event query
The akka-persistence-jdbc back end interprets offsets differently: it
starts the stream from the event _with_ the provided offset rather than
starting after that event. This causes JDBC read side processors that
restart to see the last event twice.

In Lagom 1.2.x, we compensated for this by incrementing the offset when
performing the query, but this was accidentally lost during some of the
changes in Lagom 1.3.0.

Rather than incrementing sequence offsets unconditionally, this
reimplemenation delegates to concrete persistence implementations the
decision for converting the last stored offset to the starting offset
for the stream.