Finished and fixed Scala documentation
* Added PubSub documentation
* Fixed all broken links
* Enabled documentation validation

When writing the PubSub documentation, I found a major type error in
persistent entities:

thenPersist[B <: Event](event: B, afterPersist: B => Unit): Persist[B]

The method above doesn't allow B to be inferred in the afterPersist method
since Scala type inference doesn't work within method parameter groups,
so you had to explicitly annotate the type of the argument to
afterPersist.  The reason why this didn't cause a problem so far is that
the command handler callback took an argument that returned
Persist[Event], rather than Persist[_ <: Event], forcing B to be Event,
and thereby allowing the type inference to use that to infer the type.
This never caused a problem because we didn't have any examples
afterPersist callbacks that took a subtype of Event. This changed when I
wrote the PubSub docs, which publish the PostPublished event.

So I've changed thenPersist to take two parameter groups, ie:

thenPersist[B <: Event](event: B)(afterPersist: B => Unit): Persist[B]

This has meant that I also needed to delete the thenPersist(event)
overload, since with it, the compiler can't work out which method you're
referring to.  Since commands nearly always need to be replied to in an
afterPersist callback, I don't think this is a problem.

Finally, I removed the type parameter of Persist, I don't think it's
needed in Scala from an API perspective, and the only place we used it
was for PersistOne, so now PersistOne declares a type parameter, but
there's no need to for Persist.