Akka serializer injection
Upgraded to Akka 2.4.16, and implemented Akka serializer injection using
it.

Since the serializer registry is now available before the actor system
is instantiated, we can use it to find a complete list of all the types
that it supports. This means we can get rid of `Jsonable`, as we can
register those types directly.

I've ensured that a Lagom application doesn't need to have a registry,
unless it uses Akka clustering.  This is done using two traits, one
called ProvidesJsonSerializerRegistry which has an optional registry
that defaults to None, this is what the Lagom application uses to get
the registry when it starts the actor system. The other is a
RequireJsonSerializerRegistry, which has a serializer registry def with
no default implementation, so it must be implemented by the end user.
This trait also extends ProvidesJsonSerializerRegistry, overriding the
optional one with the required one. Finally, ClusterComponents extends
RequireJsonSerializerRegistry, so if a user mixes ClusterComponents
(which is extended by PersistenceComponents and so on) into their
application, they will be forced to provide one.

I renamed some of the types, to make it clear to users that these
serializers are just for JSON - this is to avoid confusion that may be
expected if someone wants to use, for example, protobufs, and they can't
see how they can register a protobuf serializer with the
SerializerRegistry.  So, SerializerRegistry => JsonSerializerRegistry,
Serializers => JsonSerializer, and Migration => JsonMigration.