Fixed lagomExternalProject
Fixes #426.

Java projects needed to have the Java dev mode dependencies added, but
there was no way in the existing API to tell if an external project was
a Java project or a Scala project, so I deprecated the existing method
and created lagomExternalJavadslProject and
lagomExternalScaladslProject methods.

Scripted tests for both the javadsl and scaladsl cases have been
created.