Document cassandra-configuration:dist command
The page http://www.lagomframework.com/documentation/1.3.x/java/ConductRSbt.html explains which sbt commands are available when adding the plugin `sbt-conductr`.

So far the command `cassandra-configuration:dist` was not listed there. Because we are using this command later on the page I thought it makes sense to list it there as well.

I’ve also modified the text of the other respective bullet points.