Changed Kafka support to lookup Kafka from the service locator
The service locator is looked up every time a connection to Kafka is
made - I think this is what we want, since generally connections to
Kafka are long lived, and in the case that a Kafka node dies and is
restarted somewhere else, typically the stream will fail, then when it
reconnects, it will need to look up from the service locator again to
find out where another node is.

I modified the Kafka tests so that it will use this.

Passing configuration explicitly is still supported, by setting
lagom.broker.kafka.service-name to an empty string.

Fixes #511.