Fix current schema detection
This now uses the standard Connection.getSchema method (available since
Java 7) but continues to fall back on vendor-specific implementations if
it returns null (which is the case for some versions of H2 and the MySQL
driver, at least).

Also passes the current schema to the getTables search, as the default
is to pass "", which is specified as: "retrieves those without a
schema". This typically means that it returns tables in the database
default schema, which might not be the current schema.