Refactored dependencies into separate file (#581)
Refactored the dependencies into a separate file. This is an initial
step before #528 is done so that reviewing that becomes simpler.

Often projects will pull out Seqs of dependencies, rather than
Setting[Seq]. I've extracted the dependencies as Setting[Seq], as this
will make it very easy to introduce more complex things in future, like
when we cross build with Scala 2.12, it will be very straight forward to
make the dependencies dynamic based on Scala version, no changes to
build.sbt will be needed, just the dependency setting in
Dependencies.scala.