Fixed deprecation warning produced by forServer macro
The forServer macro is producing AST that uses a deprecated API, which
results in a deprecation warning in the users compilation.  This fixes
that. It also improves scaladocs and deprecation warning descriptions,
and adds the multiple service descriptors deprecation to the 1.4
migration guide.

This will need to be backported to 1.3.x (excluding the migration guide
changes).