Support well known ports in the service registry (#707)
This change allows unmanaged services to be registered using http and
https URLs that don't specify the port. In addition, it also allows URLs
which don't specify a port and for which no well known port can be
calculated, but only when no ACLs are registered in conjunction with
that service (since the port is only needed by the service gateway to
proxy requests, and so if there's no ACLs, the service gateway doesn't
need it).