Use new Cassandra launcher
Akka's CassandraLauncher now forks. It doesn't make sense for us to fork
the launcher itself, since it's just going to fork after that, so this
changes things to run the launcher in process. We're also not using the
launcher itself due to some customisations that we needed, including the
ability to specify custom JVM options.