Kafka service name override
Add the possibility to override the Kafka service name with the environment variable `KAFKA_SERVICE_NAME`. This is a common scenario in the context of ConductR on DC/OS on which the service name needs to be overwritten to `_node-0._tcp.kafka.mesos`.

As descrbed at https://tech-hub.lightbend.com/guides/dcos-microservices, the default approach within ConductR is to provide a custom bundle configuration in which the environment variable `KAFKA_SERVICE_NAME` is set and therefore the Kafka service name is overwritten.