Update jboss-logging to 3.3.0.Final
This is the version that's used everywhere in Lagom other than in the
transitive dependency of Cassandra. This change forces it to use that
version in Cassandra, too.

This fixes a license issue: the version used by Cassandra by default is
licensed under the LGPL. This uses a version licensed under Apache v2.