Force the JVM to halt in the last-resort exit
System.exit (which calls Runtime.getRuntime().exit) will block if the
runtime is currently in the process of running shutdown hooks. This
means that if the shutdown hooks deadlock, this will too.

Since the purpose of this is to force shutdown in the case where the
actor system is not stopping cleanly in a reasonable amount of time,
this commit changes the behavior to call Runtime.halt instead. This
method exits the JVM immediately, without running shutdown hooks or
otherwise cleaning up.