Added child persistent entity
This allows a persistent entity to run as a companion to another actor.
This allows the following use cases:

* An actor running as a persistent process manager can use the entity to
store its state without having to communicate with another node in the
Akka cluster.
* When you have a use case where you need to do something that is far
beyond the scope of a persistent entity (such as managing some non event
sourced data store), and the entities for that datasource can only be
acted on one at a time, so you're using cluster sharding to distribute
that processing, and you do want to emit persistent events as well for
these entities.

By running the entity as a child entity, you can cut down on network
communication and decrease your surface area for failure to impact you.