Complementary event log feature
This shows how a complementary event log feature could be offered. It
writes events directly to the akka-persistence-jdbc event log, but in a
user supplied transaction, allowing them to be consumed by Lagom's
existing read side processors and topic publishers.

So far, only the minimum for supporting this in JPA has been
implemented.

Furthermore, these changes are dependent on
https://github.com/dnvriend/akka-persistence-jdbc/pull/180, the primary
reason being that without that change, Akka persistence JDBC reads the
sequence number, which this tries to generate in a single database call
by doing a subselect, from a serialized form of an Akka class, rather
than from the table row.