Stop application resources from being returned twice from getResources (#1347)
* Stop application resources from being returned twice from getResources

The delegating classloader is a little odd - it makes the application
classloader resources available to libraries which are in the parent
classloader. However, this means when it loads resources, the resources
it loads appear twice on the classpath because it both loads its own,
and the parent, which includes its own. This ensures that it doesn't
include its own resources in getResources, that it only gets the parent
resources.

* Reformat

* Preserve resource ordering

Optimize the returned enumeration to avoid copying