Refactor persistence test configuration (#1428)
This is to ease the transition to Play 2.7.

Extracted from #1405.

## Purpose

Changes invocations of `Context.apply` to `ApplicationLoader.createContext`. This requires changing the configuration passed to it from a `play.api.Configuration` instance to a `Map[String, AnyRef]`, which triggered a larger refactoring of the hardcoded configuration data used in `ServiceTest`.

## Background Context

Right now, test configuration is spread around in a few different places. This has resulted in an awkward dependency from TestKit to the Cassandra Persistence implementation. To simplify the refactoring, this was consolidated to make TestKit more self-contained.

## References

https://github.com/lagom/lagom/pull/1405#discussion_r200846049