Adds SSL support for lagom's testkit (#1599)
This is related to #1595 but not a fix.

This is a minimal change to make ssl available in lagom's `TestService` that unblocks testing gRPC. The UX for gRPC users is not great but at least this works. Needs massaging and improving.

At the moment I keps the scope scala-only as we discuss and improve the approach.