Fix broken external links (#1793)
Also adds external link validation to the nightly build.

It shouldn't be part of the normal pull request validation since external links can change at any time, and shouldn't cause unrelated pull requests to fail. Note that the link validator considers redirects to be failures as well as 404s or other errors.

## References

I noticed these while working on #1792, and this duplicates some of the changes in there. Whichever one is merged first will cause conflicts with the other, but I can resolve those.


Co-authored-by: Ignasi Marimon-Clos <ignasi@lightbend.com>