Tolerate multi-jvm failures when running tests on CI
We still have some flaky tests for persistence-cassandra-*, as documented
in #2080. Those are the most common flaky tests so we can ignore them for
now when running on Travis or Circle. They won't be ignored locally, however.

We also won't ignore failures on cron/scheduled builds since they are a
safe guard and don't disrupt the contribution flow.