Revert AbstractPersistentEntityRegistry's binary API
This reverts the material change in #2370, now that it's fixed upstream
with Akka 2.6.0-RC2.

The reason for to revert this is #2414:
CouchbasePersistentEntityRegistry, in the lagom-persistence-couchbase
artifact, which is in the akka-persistence-couchbase repo, depends on
AbstractPersistentEntityRegistry's binary API.  So the change in #2370
broke that.

Sadly this means that eventsByTagQuery goes back to being lazily
evaluated as well as "optional" (but not really).  However, those were
side-effect improvements that we can try and re-introduce later, bearing
in mind this regression.

Fixes #2414