Run scripted tests targeting sbt 1.3
The key `sbtVersion in pluginCrossBuild` defines the target sbt version
and is the default for `scriptedSbt`, which is the key used when running
scripted.  So by leaving the former set to 1.2.8 (in defineSbtVersion at
the top of build.sbt) but setting scriptedSbt we test that the
1.2.8-targetting build runs on sbt 1.3.

To mitigate the cost on CI times, this will only run in cron builds.